+++ R/as_tibble.R
@@ -116,7 +116,8 @@ as_tibble.list <- function(x, validate = TRUE, ..., .rows = NULL,
   }
 
   x <- set_repaired_names(x, .name_repair)
-  x <- check_valid_cols(x)
+  check_valid_cols(x)
+  x[] <- map(x, strip_dim)
   recycle_columns(x, .rows)
 }
 
@@ -133,7 +134,7 @@ check_valid_cols <- function(x) {
     abort(error_time_column_must_be_posixct(names_x[posixlt]))
   }
 
-  x
+  invisible(x)
 }
 
 is_1d_or_2d <- function(x) {
@@ -147,7 +148,7 @@ recycle_columns <- function(x, .rows) {
 
   # Shortcut if all columns have the requested or implied length
   different_len <- which(lengths != nrow)
-  if (is_empty(different_len)) return(new_valid_tibble(x, nrow, subclass = NULL))
+  if (is_empty(different_len)) return(new_tibble(x, nrow = nrow, subclass = NULL))
 
   if (any(lengths[different_len] != 1)) {
     abort(error_inconsistent_cols(.rows, names(x), lengths, "`.rows` argument"))
@@ -158,7 +159,7 @@ recycle_columns <- function(x, .rows) {
     x[short] <- expand_vecs(x[short], nrow)
   }
 
-  new_valid_tibble(x, nrow, subclass = NULL)
+  new_tibble(x, nrow = nrow, subclass = NULL)
 }
 
 guess_nrow <- function(lengths, .rows) {